+++ src/models/Settings.php
@@ -63,6 +63,7 @@ class Settings extends Model
     public bool $outputDefaultCss = true;
     public bool $outputDefaultJs = true;
     public ?string $assetBaseUrl = null;
+    public ?string $assetBasePath = null;
 
     // Giphy
     public bool $giphyEnabled = false;
@@ -250,6 +251,11 @@ class Settings extends Model
     public function getAssetBaseUrl(): ?string
     {
         return App::parseEnv($this->assetBaseUrl);
+    }
+
+    public function getAssetBasePath(): ?string
+    {
+        return App::parseEnv($this->assetBasePath);
     }
 
     public function getRecaptchaSecret()